lost_and_found/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.*
asyncpg
psycopg2-binary
alembic
pydantic
python-jose
passlib[bcrypt]
httpx
pytest
pytest-asyncio
pytest-cov
reportlab
pillow
numpy
python-dotenv
asgi_lifespan
